projects/cmdb-servicenow/implementation/requirements.txt
# Python dependencies for CMDB ServiceNow Discovery
pyVmomi==7.0.3
kubernetes==29.0.0
pysnmp==4.4.12
paramiko==3.4.0
requests==2.31.0
pyyaml==6.0.1
schedule==1.2.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-mock==3.12.0
responses==0.24.1

# Dev
black==23.12.1
flake8==6.1.0
mypy==1.8.0